MB-32591: Trigger expelling to occur before cursor dropping
Currently checkpoint expelling is applied every time a vbucket is
visited via CheckpointVisitor::visitBucket.  This aggressive approach
means that items can be expelled from a checkpoint even if memory usage
is below the low watermark.  If a new DCP producer DCP stream creates a
cursor to send items to a consumer there is an increased chance that
items will have to be retrieved via backfilling, where previously they
would have been sent directly from the checkpoint.  Consequently
over expelling could unnecessarily degrade performance.

Therefore this patch changes checkpoint expelling to be triggered only
when ClosedUnrefCheckpointRemoverTask is run and the existing criteria
for cursor droping is satisified.

We first apply checkpoint expelling, then if the memory target is still
not met we perform cursor dropping.  The reason for prioritising
expelling over cursor dropping is that it is considered less disruptive.

Change-Id: I92cbf31390e1cd88f5f3c64cca05fa348441ec5b
Reviewed-on: http://review.couchbase.org/108904
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>