MB-34334 [SR]: Byteswap DcpSeqnoAcknowledgedPayload when assigned
DcpSeqnoAcknowledgedPayload.prepared_seqno is byteswapped when read,
but not when initially written. This is incorrect, as it is supposed
to be stored in network order.

However in normal usage this ends up benign, as:

a) SeqnoAcknowledgement (which has DcpSeqnoAcknowledgedPayload as a
   member) ends up double-byteswapping on send from the consumer, and

b) When the producer recieves (dcp_seqno_acknowledged_executor) it
just performs type-punning via a DcpSeqnoAcknowledgedPayload pointer
and hence never invokes the constructor.

Having said all that, add the correct byteswap to (hopefully) avoid
any confusion, and add some sanity checks to the ActiveDM that we
don't get seqno_ACKs in excess of the highest tracked seqno.

Change-Id: Id7a24f4e506fb77ed4e9f26675720c6091355ad9
Reviewed-on: http://review.couchbase.org/109726
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>