MB-33332: Fix SyncRep Ephemeral tests failing due to merge conflict
Two issues here:
1) The epehemeral Abort code did not consider the new committed state
PreparedMaybeVisible. Fix the Abort code to use the isPending check.
2) Not sure what happened to the CommitExisting_RangeRead test but
the test was invalid in that if we do any writes whilst there is a
range read then we will never attempt to write anything in that range.
Fix the test by correcting the logic to what was intended, grabbing a
range read covering an already committed value and then do another
commit.

Change-Id: I045973bcdcecc68530ed7655de7beda4afe40266
Reviewed-on: http://review.couchbase.org/109717
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>