MB-34017 [SR]: Include the prepared-seqno in Commit queued items (re-apply)
(Re-apply after fixing merge conficts with
http://review.couchbase.org/#/c/109724/)

When we complete a Prepare, we queue a Commit or Abort item into the
CheckpointManager.
At persistence of Commit/Abort, we don't know what is the Prepare
associated with that Commit/Abort.
While that information has never been necessary so far, now it is for
persisting the High Completed Seqno to disk.

This patch ensures that both the Commit and Abort queued items in CM
carry the seqno of the associated Prepare (ie, the prepared-seqno).

In follow-up patches, the prepared-seqno will be used at persistence for
computing the HCS and writing it to disk.

Change-Id: I21012f876d5afe5e40aa0783b2e5fbbc4c76fd8e
Reviewed-on: http://review.couchbase.org/109989
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>