MB-34017 [SR]: Include the prepared-seqno in Commit queued items
When we complete a Prepare, we queue a Commit or Abort item into the
CheckpointManager.
At persistence of Commit/Abort, we don't know what is the Prepare
associated with that Commit/Abort.
While that information has never been necessary so far, now it is for
persisting the High Completed Seqno to disk.

This patch ensures that both the Commit and Abort queued items in CM
carry the seqno of the associated Prepare (ie, the prepared-seqno).

In follow-up patches, the prepared-seqno will be used at persistence for
computing the HCS and writing it to disk.

Change-Id: Idb1473c6b89dcc3b7e401c616cc2d785331b4d78
Reviewed-on: http://review.couchbase.org/109718
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>