MB-33332: Rename OSV::deletedTime to OSV::completedOrDeletedTime
To purge completed prepares only after the metadata purge interval
we need to know when the prepare was completed (committed or aborted).
We do similar for deleted items, we set a deleted time on deletion
of the OSV and use this time in the HTTombstonePurger to determine if
we should purge the metadata (tombstone) for the delete. We can simply
re-use the field for completed time. Rename the functions to something
more appropriate. They will be used in a later change.

Change-Id: Ia37b3a6bba9f3945280ed9c1dc2568998cc0e122
Reviewed-on: http://review.couchbase.org/109839
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>