MB-32591: Enable checkpoint expelling
This patch enables the functionality to expel items from the oldest
checkpoint that still has cursors in it.  This can include the open
checkpoint.

Items are expelled items from checkpoints to reduce memory usage.  It is
applied before cursor dropping is attempted.

Expelling is achieved by identifying the oldest checkpoint that has
cursors still in it.  Then identifying where the first cursor is
located.  For example in the diagram below it is seqno 1004.

                  low                                high
           dummy seqno                              seqno
             |     |                                  |
             |     |                                  |
            \./   \./                                \./
         --------------------------------------------------
         | 1000 | 1001 | 1002 | 1003 | 1004 | 1005 | 1006 |
         --------------------------------------------------
                                       /.\           /.\
                                        |             |
                                        |             |
                                     cursor 1      cursor 2

It first swaps the dummy item with the item pointed to by cursor 1.
It then expels items from the start of the checkpoint upto where the
dummy item is newly located.  This means expelling 1004, 1001, 1002 and
1003.

After the expel operation the checkpoint in memory is follows.

                 low     high
          dummy seqno   seqno
            |      |      |
            |      |      |
           \./    \./    \./
         ---------------------
          1000 | 1005 | 1006 |
         ---------------------
           /.\           /.\
            |             |
            |             |
         cursor 1      cursor 2

Change-Id: Ie197648d0035510a5728ce0c4a06f279a6968e76
Reviewed-on: http://review.couchbase.org/109890
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>