MB-34393: Subdoc: Correctly handle NOT_STORED race when adding doc
During a subdoc mutation, if the user specified doc_flag::Add, and the
key didn't exist at the start of the command processing, but did exist
at the point we finally store the new value (i.e. another connection
added the key just before us), then the engine update fails with
ENGINE_NOT_STORED (as expected).

However, this status isn't correctly handled by the subdocument state
machine, resulting in an exception being thrown and the user
connection being terminated:

    2019-05-30T16:14:39.675236+01:00 WARNING 47: exception occurred in runloop during packet execution.
    <cut>
    conn_execute: Should leave conn_execute for !EWOULDBLOCK

Issue is in checking the status code from bucket_store in
subdoc_update() - if NOT_STORED and we _don't_ remap to KEY_EEXISTS
(for a retry), return a response to the client.

Change-Id: I14e93a2397e0f955b4de1cf99cf62a3d7d905cb9
Reviewed-on: http://review.couchbase.org/109909
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>