MB-34329: reset snapshot range to be a valid range
The original code at fault here (resetSnapshotRange) was added
as a fix for MB-14388. The ep_testsuite (dcp) unit test
test_failover_scenario_with_dcp does indeed trigger the same
failure as seen in the MB, revert the entire resetSnapshotRange
function and the test fails. However with the code removed
the failure cannot be observed in system testing. My
conclusion is that the test is bogus, there is no obvious takeover
occurring in the logs of the original MB.

The conclusion is that it is not safe to remove the resetSnapshotRange
function completely, but it should never set the range to be +1 of
the current seqno, so that is the change to make, simplify the function
and generate a legal range.

Change-Id: I1243c5cc4239fe285e0d41c2ea7957a445aacb75
Reviewed-on: http://review.couchbase.org/109983
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>