MB-33332: Add PrepareAborted and PrepareCompleted committed states
For Ephemeral we need to keep completed prepares in the HashTable for
DCP snapshot/seqlist consistency. We need to be able to determine if any
given prepare has been completed. We could do this using a
HighCompletedSeqno, however, this does not work for the abort case as
we modify the aborted prepare to give it a new seqno higher than the
original prepare on replicas. We also need to be able to distinguish
between committed and aborted completed StoredValues so that we can send
DCP_ABORT as required.

The PrepareAborted state will also be used to
simplify the queueAbort path so we should add the new state required to
StoredValue instead of OrderedStoredValue.

Add new PrepareAborted and PrepareCommitted states to CommittedState in
StoredValue.

Change-Id: I5a048d3155750a87a807e97761cd26bbcaf3a7c4
Reviewed-on: http://review.couchbase.org/109985
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>