MB-10291: Add support for OutOfOrder execution
If a command blocks as part of its execution pattern (returns
ENGINE_EWOULDBLOCK) and there is more commands available in
the input pipe we might start execution of the next command
if the current command allows for that (only whitelisted
commands which we know don't have any side effects).

The current version allows a connection to execute up to 32
commands "in parallel".

Change-Id: I2978d51ab63bbe922299c7c5a381c06e954dfd6e
Reviewed-on: http://review.couchbase.org/109715
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>