MB-34437: Fix lock order inversion at Consumer::seqnoAck/closeAllStreams
For SyncRep we added seqno acking that does the following locking:
1) Lock vbConns[x] - against modification of the vbConns list
2) Lock PassiveStreamMap
4) Unlock PassiveStreamMap
5) Unlock vbConns[x]

At DcpConsumer::closeAllStreams we do the following:
1) Lock PassiveStreamMap
2) Lock vbConns[x]
3) Remove from vbConns[x]
4) Unlock vbConns[x]
6) Unlock PassiveStreamMap

We have a similar pattern in the Producer however this is solved by
deferring locking to StreamsMap objects (always inside vbConns locks)
and using folly's AtomicHashMap instead of our own. It isn't trivial
to migrate the Consumer to folly's AtomicHashMap (from past experience).

Fix this for now by unlock vbConns[x] early and relying on the ownership
given to use by the locked shared_ptr. This also relies on us having only
one Consumer per vBucket per node. In the future when we look at MB-33739
 again we will likely need to modify the Consumer's map.

Change-Id: I919dc4649bd6f364663c4691334dc78430b9142b
Reviewed-on: http://review.couchbase.org/110061
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>