MB-34409: Fix TSAN data race in cluster_tests
During shutdown we had one thread closing the socket descriptor
to notify the thread reading and writing in the replication
that it should terminate its libevent loop.

TSAN detects that this is a data race since we've got multiple
threads operating on the same socket descriptor.

Change the logic to work like memcached; use a socketpair and
send something over the pipe to trigger a read event in the
other end and break out of the event loop.

Change-Id: I5317582a7f5d69d178c600594ef60b44d66e5cd4
Reviewed-on: http://review.couchbase.org/110088
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>