MB-34412: Hoist redirection for new prepare to processSet
When we attempt to update a StoredValue that is not a prepare with
a StoredValue that is a prepare, we redirect the updateStoredValue
to addNewStoredValue. This is common logic that we will need to be
able to skip over when we wish to allow replicas to accept multiple
prepares in a row (they should mutate the existing one instead) so
pull the logic into vBucket::processSet.

Change-Id: Iaf451e4d22fc48cbc6921fc0880c3294278f2ea3
Reviewed-on: http://review.couchbase.org/110235
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>