MB-34412: Accept duplicate prepares at replica
For a given window when a consumer resumes from a previous point it
may receive two prepares in a row if the producer de-deuplicates the
DCP Commit message. In this case, we should overwrite the existing
prepare. The given window for which this is valid is for any prepare
that the consumer has already received between the HighCompletedSeqno
and the HighPreparedSeqno.

Change-Id: I683b12e9930c8fedd6290b5c59c298d51e7fe051
Reviewed-on: http://review.couchbase.org/110367
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>