MB-34687: Add prepareSeqno to DCPCommit
We need the prepared seqno in the case where are replica receives two
commits in a row. In this case it may not have a prepare in the HashTable
from which to get the prepareSeqno. If it does, the seqno will be wrong.

There exists two issues currently where we still allow 0 prepareSeqnos
which causes the sending of a Mutation instead of a Commit. These are
due to disk backfill doing a CacheLookup and creating an Item from a
StoredValue (which does not have a prepareSeqno so defaults to 0) and
Ephemeral not setting the prepareSeqno on the OrderedStoredValue. These
will be fixed in the following commit as they would cause a unit test
failure that requires a fix that should be highlighted in a separe
commit for MB-34542.

Change-Id: Ifdcdb1b68d12270587267912e307693c43edd705
Reviewed-on: http://review.couchbase.org/111005
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>