MB-34473: Add prepareSeqno to DCPAbort
We need the prepared seqno in the case where are replica receives two
aborts in a row. In this case it may not have a prepare in the HashTable
from which to get the prepareSeqno. We will use the prepareSeqno to
validate if the second DCPAbort can be safely ignored. This is only the
case when resuming a stream and the
highSeqno < prepareSeqno < streamStartSeqno. If the seqno does not fall
in these bounds, we want to throw an exception as the producer may have
done something wrong and we may be breaking durability constraints.

Change-Id: I40f27e7601baac00c06de9ad8176e796a5995636
Reviewed-on: http://review.couchbase.org/110249
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>