MB-34542: Do not send commit for de-duped prepares to replica
In the case where we de-dupe a prepare on the resumption of a
stream, a consumer would recieve a commit message for which it
would not have a corresponding prepare. This is an issue as
we do not have the corresponding value to commit. In this case,
we will instead send the consumer a normal mutation. This carries
all of the information required to "commit" the value and will not
be blocked with InFlightSyncWrite as logically the consumer would not
have a prepare pending.

Change-Id: I8827a596dfc5cc326782f735d9d0b82b0fbe4e5d
Reviewed-on: http://review.couchbase.org/110369
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>