MB-34495 [SR]: Ephemeral: Don't create duplicate Prepared items in replica HT
For Ephemeral buckets, when processing DCP_COMMIT messages there could
already be a completed prepared_sync_write OSV in the HashTable. If
so, then this existing prepare object should be re-used; we
_shouldn't_ create a new one and add that into the HashTable.

Bug manifests as DCP_COMMIT failing with the following exception:

    GSL: Precondition failure at kv_engine/engines/ep/src/hash_table.cc: 296

Which refers to the following expectation in HashTable::findInner:

    if (v->hasKey(key)) {
        if (v->isPending() || v->isCompleted()) {
            Expects(!foundPend);   <<<< HERE
            foundPend = v;
        }
        ...

Fix by chcking for an existing commit by using
HashTable::findForSyncWrite() in VBucket::setWithMeta() when the
incoming item has durability requirements.

Change-Id: I05cba04f6677ab168e28975dcd868c49d3ce9a40
Reviewed-on: http://review.couchbase.org/110303
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>