MB-34507: Reset datatype to XATTR when pruning system XATTRs
When pruning system XATTRs during expiry, if the input document was
compressed then we fail to clear the SNAPPY flag in the datatype -
even though the pruned XATTRs /will/ be uncompressed. This results in
subsequent attempts to manipulate the document body failing as
KV-Engine will think it needs to decompress the body, however that
will fail as Snappy will (correctly) return the data isn't Snappy
compressed.

For example, this manifests when attempting to modify via subdoc:

    WARNING <4 ERROR: Failed to determine inflated body size. Key: '<ud>docid</ud>' may have an incorrect datatype of COMPRESSED_JSON.

Fix by reseting the datatype to XATTRs if the pre-expiry callback
returns a non-empty response.

Change-Id: I73db03c329da79ba5261f8af185854324c1c54c3
Reviewed-on: http://review.couchbase.org/110318
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Daniel Owen <owend@couchbase.com>