MB-34542: Do not send commit for de-duped prepares to replica
In the case where we de-dupe a prepare on the resumption of a
stream, a consumer would recieve a commit message for which it
would not have a corresponding prepare. This is an issue as
we do not have the corresponding value to commit. In this case,
we will instead send the consumer a normal mutation. This carries
all of the information required to "commit" the value. The consumer
must also be aware that the produce can do this and allow the overwrite
of an existing prepare with a set when the existing prepare has seqno
less than the stream start.

Change-Id: Id0f1009baec4710634fdbd3ff63167c7703f377e
Reviewed-on: http://review.couchbase.org/110583
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>