MB-34562: Don't throw duplicate mutation if receiving first snapshot
If we are receving our initial backfill from disk then we will not
receive any snapshot markers denoting the end of a checkpoint until
the entire on disk snapshot has been received and we switch over to
InMemory. As such, we can't guarantee that durable ops are in separate
checkpoints at replica in this case.

Change-Id: I3f0b364723a58fa1e36e8ce250852e6a59ff0d85
Reviewed-on: http://review.couchbase.org/110600
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>