MB-34569: Do not throw at seqnoAck if stream not found
If the flusher attempts to do a seqnoAck (for persistence
durability levels) then it may find that a stream no longer exists
and the exception that is thrown will not be caught. This will cause
memcached to be aborted which is not ideal. Instead of aborting
memcached in this case, just log the error as the stream could have
gone away for a valid reason/handled exception in runloop and don't
attempt to process the seqnoAck.

Change-Id: I0bfe2f51357981f2c16aa02e075bb0750e362aa9
Reviewed-on: http://review.couchbase.org/110711
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>