MB-34380: Reset the vbucket state when deleting a vbucket
The MB highlighted that the absence of any clearing/reset of
the KVStore vbucket state duing vbucket deletion could lead
to a subsequent vbucket create skipping the persistence of
the first vbucket state-change, leading to no vbucket file
on disk. This manfested itself in a no-operation situation
as handled exceptions and warnings in the log files from
statistics requests which wish to peek at each vbucket file.

In the fix we now reset the vbucket state as part of the
pre-existing prepareToDelete function.

To note:

prepareToDelete is called whilst the vbucket exists in
the vbMap, however the vbucket-map 'element' lock is held
to ensure we serialise with any subsequent create.

The flusher could be holding the to-be-deleted vb (because of
shared_ptr) and could flush the now reset state, however the
revision of file that state was flushed to is to be deleted.

Change-Id: I56fcadc7213bc56a791d7713d399d4eda442f56b
Reviewed-on: http://review.couchbase.org/110747
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>