MB-34314: Add HCS and HPS to RollbackResult
During a rollback we cannot set the HCS and HPS correctly if we
rollback all in-flight SyncWrites even though we do have the
information required to do so (on disk in the db header that we are
rolling back to). Att the HCS and HPS to the rollback result so that
we can correctly set the HCS and HPS of the PassiveDurabilityMonitor
post rollback.

Change-Id: I15dfa5fb059b8d5b0ea6c75d27ffac2d50092c1b
Reviewed-on: http://review.couchbase.org/110805
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>