Refactor: Use persisted vbucket_state when flushing
Currently we base any vbucket_state changes on the in-memory
vbucket_state when flushing. This is not just error prone, but it's
also problematic when we want to persist the HighPreparedSeqno (HPS).
Currently this will reset the HighCompletedSeqno (HCS) to 0 as we leave
it defaulted (0) when return the in-memory vbucket_state. This then
causes the update of the on disk vbucket_state to overwrite the
HighCompletedSeqno.

We should instead base any changes to the vbucket_state on the on disk
vbucket_state so that we do not overwrite unchanged values
unintentionally.

Change-Id: Idee8ab4d402d554ef91863c19c148ff20953c1fe
Reviewed-on: http://review.couchbase.org/110819
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>