MB-34565: Move RollbackResult to new file
RollbackResult should also logically contain the prepares that we need
to re-add to the PassiveDM. This requires it to own a container of
queued_items which requires it to have the full definition of an Item.
Move RollbackResult to a new file so that we do not have a circular
reference between ep_types.h and item.h (via storeddockey.h) and do not
have to include item.h in everything including ep_types.h.

Change-Id: I400e6c027b4bf40383e4c3cf918941deb0ccfb7b
Reviewed-on: http://review.couchbase.org/110931
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>