MB-34676: Persist the prepareSeqno of commits and aborts
This requires add a prepareSeqno to:
 - Couchstore MetaDataV3 (used for prepares, commits and aborts)
 - RocksDB MetaData (used for all values)
 - Magma MetaData (used for all values)
 - OrderedStoredValue (used for all values)

We need the prepareSeqno to be correct when backfilling from disk
to keep our sanity checks on replica and avoid sending mutations
instead of commits wherever possible. Without this, we would need to
send a mutation instead of a commit for every backfilled commit. The
prepareSeqno will also be used for aborts to simplify some error
checking on the replica when receiving a backfilled abort.

Change-Id: If203af661b976ca3145eec1873ae9f30bce80c6c
Reviewed-on: http://review.couchbase.org/110998
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>