MB-34516: Make PassiveDM correctly handle prepares with deduped commits
Given an existing tracked Prepare for key1, if we are receiving a disk
backfill, if the "matching" Commit for key1 has been deduped the PDM
might encounter a Commit for key2 before the new Commit for key1.

Handle this by allowing the PassiveDM to "skip ahead" - implicitly
complete any prepares it expected to see be committed *before* the one
it actually received. The active only sends out of order commits in the
dedupe case, so we know we will not see the commit for these prepares.

Importantly, we know we *will* see a new prepare (which caused the
dedupe) for the key before any further Commit for it.

Change-Id: Icde4a9e991c218e02d30a7966b5c30b7b8a8b498
Reviewed-on: http://review.couchbase.org/111698
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>