MB-33463: Add support for synchronous compaction for magma
Compaction in magma is asynchronous. Its triggered by 4 conditions:
* Level compaction
* Expiry compaction
  In magma, a histogram of when items will expire is maintained.
  Periodically, magma will trigger a compaction to visit those
  sstables which have expired items to have them removed.
* Dropped collections removal
* Tests
  For tests which call compactDB, we will support synchronous
  compaction.

For DP, will will support level compaction and synchronous
compaction.

The StorageProperties::ConcurrentWriteCompact is set to No. This is a
bit misleading in that magma does support concurrent compaction and
writes, it just doesn't support the ability to update the datastore
from a thread other than the bg_flusher thread. This will be
addressed when the compaction_ctx callback is implemented.

Change-Id: Ib57fafefee008da817896a578ded5711aa9079bd
Reviewed-on: http://review.couchbase.org/110979
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>