MB-34687/MB-34542: Always send non-zero prepareSeqno
Currently we may send a prepareSeqno of 0 in a commit if we are
backfilling from disk and the StoredValue for the commit exists in the
HashTable or if we are backfilling from an Ephemeral bucket.

Correcting this causes us to stop sending a Mutation instead of a Commit
because we know have the correct prepareSeqno. The condition used to
govern this though is wrong. We should not be checking against the
stream start_seqno_ (requested by the replica); we should instead be
checking that the prepareSeqno is larger than the seqno of the first item
sent in the stream. This also corrects MB-34542.

Change-Id: I8c13de37317a508aebf9a8a0baf673366b446432
Reviewed-on: http://review.couchbase.org/111025
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>