MB-34634: Test that pre(k1)/commit(k1) doesn't crash KV
Add a dcp test case which triggered the issue seen in MB-34634.
The issue was a violation of the Monotonic active_stream
lastReadSeqno member variable where following the steps shown
in the test the stream tries to send back the seqno the
consumer already has.

The issue itself was addressed by changes under MB-34685.

Change-Id: I94480b3400aead46a8ec2fca0a2e3477297b69e3
Reviewed-on: http://review.couchbase.org/111204
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>