MB-34685: Allow prepares and commits in same checkpoint
It's problematic to split prepares and commits into separate
checkpoints as a replica backfilling from disk must special case
this and create it's own checkpoints. This leads to issues such as
MB-34634.

Store an additional Prepare or Committed flag in the checkpoint
manager's keyIndex to allow prepares and commits to exist in the
same checkpoint.

Change-Id: I528830a4625cf90b134d15ff78a6fd76b746d2c8
Reviewed-on: http://review.couchbase.org/111148
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>