MB-34542: Allow overwrite of prepare when streaming from disk
We need to always allow the overwrite of a prepare with a mutation
when a replica/pending vBucket is streaming from disk. The active
node will send mutations instead of commits if the consumer node
has missed a prepare so the consumer must unconditionally accept
this mutation and overwrite (if it has one) it's existing prepare.

Change-Id: I43ed9c8ec026e4cc218b0bb4408b1b63028f30dd
Reviewed-on: http://review.couchbase.org/111212
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>