MB-34737: Remove overwritten Prepares from the PassiveDM
Replicas may encounter two prepares for the same key without seeing a
commit between them. If the replica disconnects after receiving a
prepare, if it reconnects and backfills, the commit/abort associated
with the prepare may have been deduped.

We allow the new prepare to overwrite the old one in the hashtable, but
we should also remove the old prepare from the PDM
Change-Id: I00916cb82b71beb19bd2c51215d4affe2a0b7c27
Reviewed-on: http://review.couchbase.org/111161
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>