MB-34664: Only update the snapshot range if we're in it
The flusher now only updates the snapshot range of the vbucket and
kvstore when it is writing items, where items are things that belong
to a snapshot. That is normal documents, system-events, prepares
etc... So if the flusher runs and only a vbucket-state change is found
or just a checkpoint start/end,  then we will not change the snapshot.

The original MB and unit-test in this commit demonstrate why this
change is important as we should not update the various places that
store what range we are in unless we are really in that range.

Below shows before and after the fix, the "Before" and "After" columns
shows the range which is stored in the VBucket and the KVStore
vbucket_state. The column "HS" shows the high-seqno (which hasn't
changed with this commit)

Event | Replica Events         | HS | Before | After
1     | snapshot, start:1 end:2| 0  | 1,2    | 0,0
2     | mutation k1 seq:1      | 1  | 1,2    | 1,2
3     | mutation k2 seq:2      | 2  | 1,2    | 1,2
4     | snapshot start:3, end:3| 2  | 3,3    | 1,2
5     | set state active       | 2  | 3,3    | 1,2

After event 5, the 'badness' is propagated to the failover table

Before: new failover entry high_seqno:3
After:  new failover entry high_seqno:2

Additionally without this commit as seen in the MB, when the flusher
runs and goes to store the set_vb_state (event 5) an exception is
thrown because the range the VBucket checkpoint manager is aware of is
still 1,2 but we try and set the start to be the value of 3 (read from
the VBucket's incorrect 3,3 state).

Change-Id: I5418139c9a8b11a7126c48ca022c79a676b4dbd5
Reviewed-on: http://review.couchbase.org/111356
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>